[DATA-MODELING] Reoredening the layers
