--- app/src/main/java/com/example/liverpooltest/MetaData.kt
@@ -1,3 +1,0 @@
-package com.example.liverpooltest
-
-class MetaData
