--- app/src/main/java/com/example/liverpooltest/PlpResults.kt
@@ -1,12 +1,0 @@
-package com.example.liverpooltest
-
-data class PlpResults(
-    val customUrlParam: CustomUrlParam,
-    val label: String,
-    val metaData: MetaData,
-    val navigation: Navigation,
-    val plpState: PlpState,
-    val records: List<Record>,
-    val refinementGroups: List<RefinementGroup>,
-    val sortOptions: List<SortOption>
-)
