+++ app/src/main/java/com/example/liverpooltest/model/CustomUrlParam.kt
@@ -1,0 +1,3 @@
+package com.example.liverpooltest.model
+
+class CustomUrlParam
