+++ app/src/main/java/com/example/liverpooltest/model/MetaData.kt
@@ -1,0 +1,3 @@
+package com.example.liverpooltest.model
+
+class MetaData
