+++ app/src/main/java/com/example/liverpooltest/model/Refinement.kt
@@ -1,4 +1,4 @@
-package com.example.liverpooltest
+package com.example.liverpooltest.model
 
 data class Refinement(
     val colorHex: String,
