+++ app/src/main/java/com/example/liverpooltest/model/RefinementGroup.kt
@@ -1,4 +1,4 @@
-package com.example.liverpooltest
+package com.example.liverpooltest.model
 
 data class RefinementGroup(
     val dimensionName: String,
