+++ app/src/main/java/com/example/liverpooltest/model/SortOption.kt
@@ -1,4 +1,4 @@
-package com.example.liverpooltest
+package com.example.liverpooltest.model
 
 data class SortOption(
     val label: String,
