+++ app/src/main/java/com/example/liverpooltest/model/data/MainModel.kt
@@ -1,4 +1,6 @@
-package com.example.liverpooltest
+package com.example.liverpooltest.model.data
+
+import com.example.liverpooltest.model.Status
 
 data class MainModel(
     val pageType: String,
